LayoutTests/fast/selectors/querySelector-id-with-multiple-elements-with-same-id-expected.txt

This test makes sure WebKit element.querySelector("#foo") returns a descendent node of the element even when there are multiple elements with the id foo.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS container.querySelectorAll("#foo").length is 1
PASS container.querySelectorAll("#foo")[0] is container.firstChild
PASS successfullyParsed is true

TEST COMPLETE
